# Edit and Problem-Solving
## Edit
- things
- readme.md file

## Problem Solving

### W3Resource

#### Object Oriented Programming

- Encapsulation (7)

+++ src/_2_problem_solving/_2_2_w3resource/_2_2_5_object_oriented_programming/_2_2_5_2_abstract_class/_2_2_5_2_2_question_409/Question409.java
@@ -41,12 +41,12 @@ class Circle extends Shape {
 
     @Override
     double calculatePerimeter() {
-        return (double) (2 * Math.PI * RADIUS);
+        return 2 * Math.PI * RADIUS;
     }
 
     @Override
     double calculateArea() {
-        return (double) (Math.PI * Math.pow(RADIUS, 2));
+        return Math.PI * Math.pow(RADIUS, 2);
     }
 }
